made server response handling logic globally available and implemented it into company sidebar
