+++ custom_figure.html
@@ -7,13 +7,14 @@
   <title>Neue Kennzahl erstellen</title>
   <link rel="stylesheet" href="custome_figure.css" />
   <link rel="stylesheet" href="sidebar.css" />
-    <script src="customFigureBuilder.js" defer></script>
+  <link rel="stylesheet" href="infobox.css">
+  <script type="module" src="customFigureBuilder.js" defer></script>
 </head>
 <body>
 <div class="content-wrapper">
   <div class="main-container">
 
-    <div class="infobox-overlay" id="customFigureBuilderInfoBox">
+    <div class="infobox-overlay" id="infoBoxOverlay">
       <div class="infobox"></div>
     </div>
