+++ custome_figure.css
@@ -133,34 +133,3 @@ body {
     margin-bottom: 1rem;
 }
 
-
-
-.infobox-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.6);
-    display: none;
-    justify-content: center;
-    align-items: center;
-    z-index: 1;
-}
-
-.infobox {
-    background-color: white;
-    padding: 20px 30px;
-    border-radius: 8px;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.2);
-    font-size: 18px;
-    text-align: center;
-}
-
-.success-message {
-    color: #25B24E;
-}
-
-.error-message {
-    color: #da1e28;
-}
