+++ index.html
@@ -7,11 +7,16 @@
   <script src="https://cdn.tailwindcss.com"></script>
   <link rel="stylesheet" href="styles.css" />
   <link rel="stylesheet" href="sidebar.css">
-  <script src="companySidebar.js" defer></script>
+  <link rel="stylesheet" href="infobox.css">
+  <script type="module" src="companySidebar.js" defer></script>
 </head>
 
 
 <body class="bg-gray-100 flex flex-col min-h-screen">
+
+<div class="infobox-overlay" id="infoBoxOverlay">
+  <div class="infobox"></div>
+</div>
 
 <div class="flex-grow flex flex-row">
   <div class="main-container flex-grow p-4">
